функции создания разметки для offers и destinations убраны в  общий файл
